refactor: remove obsolete HTTP routes & DDD: ModelRepo
