--- aiengine/runtime/neurdbrt/app/routes/__init__.py
@@ -1,5 +1,0 @@
-from .context import *
-from .finetune import *
-from .inference import *
-from .routes import *
-from .train import *
