--- aiengine/runtime/neurdbrt/connection/__init__.py
@@ -1,1 +1,0 @@
-from .nr import NeurDBModelHandler
